chore(schema): Reorganize category utils
