+++ modules/mvt/src/lib/mapbox-vector-tile/vector-tile-layer.ts
@@ -81,18 +81,18 @@ function readValueMessage(pbf: Protobuf) {
       tag === 1
         ? pbf.readString()
         : tag === 2
-        ? pbf.readFloat()
-        : tag === 3
-        ? pbf.readDouble()
-        : tag === 4
-        ? pbf.readVarint64()
-        : tag === 5
-        ? pbf.readVarint()
-        : tag === 6
-        ? pbf.readSVarint()
-        : tag === 7
-        ? pbf.readBoolean()
-        : null;
+          ? pbf.readFloat()
+          : tag === 3
+            ? pbf.readDouble()
+            : tag === 4
+              ? pbf.readVarint64()
+              : tag === 5
+                ? pbf.readVarint()
+                : tag === 6
+                  ? pbf.readSVarint()
+                  : tag === 7
+                    ? pbf.readBoolean()
+                    : null;
   }
 
   return value;
