+++ modules/schema/src/category/mesh/mesh-utils.ts
@@ -1,7 +1,7 @@
 // Mesh category utilities
 // TODO - move to mesh category module, or to math.gl/geometry module
 import {TypedArray} from '../../types';
-import {MeshAttributes} from '../mesh';
+import {MeshAttributes} from './mesh-types';
 
 type TypedArrays = {[key: string]: TypedArray};
