+++ modules/schema/src/lib/batches/table-batch-builder.ts
@@ -1,5 +1,5 @@
 import type {Schema} from '../schema';
-import type {TableBatch} from '../../category/table';
+import type {TableBatch} from '../../category/table/table-types';
 import type {TableBatchAggregator, TableBatchConstructor} from './table-batch-aggregator';
 import BaseTableBatchAggregator from './base-table-batch-aggregator';
 import RowTableBatchAggregator from './row-table-batch-aggregator';
